Refactor showMcf into hook
